--- a/urlscan/bin/komand_urlscan
+++ b/urlscan/bin/komand_urlscan
@@ -4,10 +4,10 @@ import komand
 from komand_urlscan import connection, actions, triggers
 
 
-Name = 'urlscan.io'
-Vendor = 'rapid7'
-Version = '2.1.3'
-Description = 'Analyze URLs for malicious indicators using the URLScan website scanner'
+Name = "urlscan.io"
+Vendor = "rapid7"
+Version = "2.1.4"
+Description = "Analyze URLs for malicious indicators using the URLScan website scanner"
 
 
 class ICONUrlscan(komand.Plugin):