--- a/urlscan/plugin.spec.yaml
+++ b/urlscan/plugin.spec.yaml
@@ -7,7 +7,7 @@ vendor: rapid7
 support: community
 status: [] 
 description: Analyze URLs for malicious indicators using the URLScan website scanner
-version: 2.1.3
+version: 2.1.4
 resources:
   source_url: https://github.com/rapid7/insightconnect-plugins/tree/master/urlscan
   license_url: https://github.com/rapid7/insightconnect-plugins/blob/master/LICENSE