--- a/urlscan/setup.py
+++ b/urlscan/setup.py
@@ -2,12 +2,12 @@
 from setuptools import setup, find_packages
 
 
-setup(name='urlscan-rapid7-plugin',
-      version='2.1.3',
-      description='Analyze URLs for malicious indicators using the URLScan website scanner',
-      author='rapid7',
-      author_email='',
-      url='',
+setup(name="urlscan-rapid7-plugin",
+      version="2.1.4",
+      description="Analyze URLs for malicious indicators using the URLScan website scanner",
+      author="rapid7",
+      author_email="",
+      url="",
       packages=find_packages(),
       install_requires=['komand'],  # Add third-party dependencies to requirements.txt, not here!
       scripts=['bin/komand_urlscan']