--- a/urlscan/tests/get_scan_results.json
+++ b/urlscan/tests/get_scan_results.json
@@ -3,7 +3,7 @@
         "action": "get_scan_results",
         "connection": {
             "api_key": {
-                "secretKey": "7016905f-c444-4610-9296-c5074b86425e"
+                "secretKey": ""
             }
         },
         "input": {