--- a/urlscan/tests/search.json
+++ b/urlscan/tests/search.json
@@ -3,7 +3,7 @@
         "action": "search",
         "connection": {
             "api_key": {
-                "secretKey": "7016905f-c444-4610-9296-c5074b86425e"
+                "secretKey": ""
             }
         },
         "input": {