--- a/urlscan/tests/submit_url_for_scan.json
+++ b/urlscan/tests/submit_url_for_scan.json
@@ -3,7 +3,7 @@
         "action": "submit_url_for_scan",
         "connection": {
             "api_key": {
-                "secretKey": null
+                "secretKey": "7016905f-c444-4610-9296-c5074b86425e"
             }
         },
         "input": {