Use input and output constants | Added "f" strings

## Proposed Changes

### Description

Describe the proposed changes:

  -

### Testing

Any testing information goes here.

## PR Requirements

Developers, verify you have completed the following items by checking them off:

### Style

Review the [style guide](https://komand.github.io/python/style.html)

- [x] For dependencies, pin [OS package](https://komand.github.io/python/style.html#dockerfile) and [Python package](https://komand.github.io/python/style.html#requirements-txt) versions
- [x] For security, set least privileged account with ``USER nobody`` in the ``Dockerfile`` when possible
- [x] For size, use the [slim SDK images](https://komand.github.io/python/sdk.html#sdk-versions) when possible: ``komand/python-3-37-slim-plugin`` and ``komand/python-3-37-plugin``
- [x] For error handling, use of [PluginException](https://komand.github.io/python/error_handling.html#plugin-exceptions) and [ConnectionTestException](https://komand.github.io/python/error_handling.html#connection-exceptions)
- [x] For logging, use [self.logger](https://komand.github.io/python/sdk.html#logging)
- [x] For docs, use [changelog style](https://komand.github.io/python/style.html#changelog)
- [x] For docs, validate markdown with ``make validate`` which calls ``mdl`` to lint ``help.md``

### Functional Checklist
- [x] Work fully completed
- [x] Functional
  - [x] Any new actions/triggers include JSON [test files](https://komand.github.io/python/style.html#tests) in the `tests/` directory created with `icon-plugin run -c sample $action > tests/$action.json`
  - [x] Tests should all pass unless it's a negative test. Negative tests have a naming convention of `tests/$action_bad.json`
  - [x] Unsuccessful tests should fail by raising an exception causing the plugin to die and an object should be returned on successful test
  - [x] Add functioning test results to PR, sanitize any output if necessary
    * Single action/trigger `icon-plugin run -T tests/example.json --debug --jq`
    * All actions/triggers shortcut `icon-plugin run -T all --debug --jq` (use PR format at end)
  - [x] Add functioning run results to PR, sanitize any output if necessary
    * Single action/trigger `icon-plugin run -R tests/example.json --debug --jq`
    * All actions/triggers shortcut `icon-plugin run -R all --debug --jq` (use PR format at end)

## Assessment
### Run

<details>

```
{
  "body": {
    "log": "rapid7/urlscan.io:2.1.4. Step name: get_scan_results\n",
    "meta": {},
    "output": {
      "scan_results": {
        "console": [],
        "cookies": [],
        "globals": [
          {
            "prop": "onformdata",
            "type": "object"
          },
          {
            "prop": "onpointerrawupdate",
            "type": "object"
          }
        ],
        "links": [],
        "requests": [
          {
            "request": {
              "documentURL": "https://a1.wpedn.com/u",
              "frameId": "62F93008F4B77C2C23BAD618EDA45481",
              "hasUserGesture": false,
              "initiator": {
                "type": "other"
              },
              "loaderId": "7041915FE2286C5C0F5B770179C5958B",
              "request": {
                "headers": {
                  "Sec-Fetch-User": "?1",
                  "Upgrade-Insecure-Requests": "1",
                  "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36"
                },
                "initialPriority": "VeryHigh",
                "method": "GET",
                "mixedContentType": "none",
                "referrerPolicy": "no-referrer-when-downgrade",
                "url": "https://a1.wpedn.com/u"
              },
              "requestId": "7041915FE2286C5C0F5B770179C5958B",
              "timestamp": 18473660.954042,
              "type": "Document",
              "wallTime": 1579976485.20375
            },
            "response": {
              "asn": {
                "asn": "16276",
                "country": "FR",
                "date": "2001-02-15",
                "description": "OVH, FR",
                "ip": "91.134.186.165",
                "name": "OVH",
                "registrar": "ripencc",
                "route": "91.134.0.0/16"
              },
              "dataLength": 585,
              "encodedDataLength": 911,
              "geoip": {
                "area": 500,
                "city": "",
                "country": "FR",
                "country_name": "France",
                "eu": "1",
                "ll": [
                  48.8582,
                  2.3387
                ],
                "metro": 0,
                "range": [
                  1535556256,
                  1535556271
                ],
                "region": "",
                "timezone": "Europe/Paris"
              },
              "hash": "532f2525d4676a66ea1bc1af53eab168e1c26a821b1726da7541dbb9eabcd1e1",
              "hashmatches": [],
              "rdns": {
                "ip": "91.134.186.165",
                "ptr": "rev.kvrmx.com"
              },
              "requestId": "7041915FE2286C5C0F5B770179C5958B",
              "response": {
                "encodedDataLength": 326,
                "fromPrefetchCache": false,
                "headers": {
                  "Connection": "keep-alive",
                  "Content-Length": "585",
                  "Content-Type": "text/html;charset=utf-8",
                  "Date": "Sat, 25 Jan 2020 18:21:25 GMT",
                  "Expires": "Thu, 01 Jan 1970 00:00:00 GMT",
                  "Server": "nginx",
                  "Set-Cookie": "JSESSIONID=node01hqw3pivwm6oz11emwq5cjuaiv138457.node0; Path=/Woodpecker",
                  "Strict-Transport-Security": "max-age=31536000"
                },
                "mimeType": "text/html",
                "protocol": "http/1.1",
                "remoteIPAddress": "91.134.186.165",
                "remotePort": 443,
                "requestHeaders": {
                  "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3",
                  "Accept-Encoding": "gzip, deflate, br",
                  "Cache-Control": "no-cache",
                  "Connection": "keep-alive",
                  "Host": "a1.wpedn.com",
                  "Pragma": "no-cache",
                  "Sec-Fetch-Mode": "navigate",
                  "Sec-Fetch-Site": "none",
                  "Sec-Fetch-User": "?1",
                  "Upgrade-Insecure-Requests": "1",
                  "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36"
                },
                "securityDetails": {
                  "certificateId": 0,
                  "certificateTransparencyCompliance": "unknown",
                  "cipher": "AES_256_GCM",
                  "issuer": "Let's Encrypt Authority X3",
                  "keyExchange": "ECDHE_RSA",
                  "keyExchangeGroup": "X25519",
                  "protocol": "TLS 1.2",
                  "sanList": [
                    "a1.wpedn.com"
                  ],
                  "signedCertificateTimestampList": [],
                  "subjectName": "a1.wpedn.com",
                  "validFrom": 1577962837,
                  "validTo": 1585738837
                },
                "securityHeaders": [
                  {
                    "name": "Strict-Transport-Security",
                    "value": "max-age=31536000"
                  }
                ],
                "securityState": "secure",
                "status": 200,
                "statusText": "OK",
                "timing": {
                  "connectEnd": 185.667,
                  "connectStart": 20.984,
                  "dnsEnd": 20.984,
                  "dnsStart": 1.591,
                  "proxyEnd": -1,
                  "proxyStart": -1,
                  "pushEnd": 0,
                  "pushStart": 0,
                  "receiveHeadersEnd": 244.978,
                  "requestTime": 18473660.954761,
                  "sendEnd": 185.971,
                  "sendStart": 185.803,
                  "sslEnd": 185.654,
                  "sslStart": 36.358,
                  "workerReady": -1,
                  "workerStart": -1
                },
                "url": "https://a1.wpedn.com/u"
              },
              "size": 585,
              "type": "Document"
            }
          },
          {
            "initiatorInfo": {
              "host": "a1.wpedn.com",
              "type": "parser",
              "url": "https://a1.wpedn.com/u"
            },
            "request": {
              "documentURL": "https://a1.wpedn.com/u",
              "frameId": "62F93008F4B77C2C23BAD618EDA45481",
              "hasUserGesture": false,
              "initiator": {
                "lineNumber": 9,
                "type": "parser",
                "url": "https://a1.wpedn.com/u"
              },
              "loaderId": "7041915FE2286C5C0F5B770179C5958B",
              "request": {
                "headers": {
                  "Referer": "https://a1.wpedn.com/u",
                  "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36"
                },
                "initialPriority": "VeryHigh",
                "method": "GET",
                "mixedContentType": "none",
                "referrerPolicy": "no-referrer-when-downgrade",
                "url": "https://a1.wpedn.com/css/fonts.css"
              },
              "requestId": "4996.2",
              "timestamp": 18473661.203155,
              "type": "Stylesheet",
              "wallTime": 1579976485.452862
            },
            "response": {
              "asn": {
                "asn": "16276",
                "country": "FR",
                "date": "2001-02-15",
                "description": "OVH, FR",
                "ip": "91.134.186.165",
                "name": "OVH",
                "registrar": "ripencc",
                "route": "91.134.0.0/16"
              },
              "dataLength": 2813,
              "encodedDataLength": 3067,
              "geoip": {
                "area": 500,
                "city": "",
                "country": "FR",
                "country_name": "France",
                "eu": "1",
                "ll": [
                  48.8582,
                  2.3387
                ],
                "metro": 0,
                "range": [
                  1535556256,
                  1535556271
                ],
                "region": "",
                "timezone": "Europe/Paris"
              },
              "hash": "fe3a16d9ff2bd64874ec75b523b14adf20fbbca652482815dd1383af41692336",
              "hashmatches": [],
              "rdns": {
                "ip": "91.134.186.165",
                "ptr": "rev.kvrmx.com"
              },
              "requestId": "4996.2",
              "response": {
                "encodedDataLength": 254,
                "fromPrefetchCache": false,
                "headers": {
                  "Accept-Ranges": "bytes",
                  "Connection": "keep-alive",
                  "Content-Length": "2813",
                  "Content-Type": "text/css",
                  "Date": "Sat, 25 Jan 2020 18:21:25 GMT",
                  "Last-Modified": "Tue, 12 Jun 2018 05:53:02 GMT",
                  "Server": "nginx",
                  "Strict-Transport-Security": "max-age=31536000"
                },
                "mimeType": "text/css",
                "protocol": "http/1.1",
                "remoteIPAddress": "91.134.186.165",
                "remotePort": 443,
                "securityDetails": {
                  "certificateId": 0,
                  "certificateTransparencyCompliance": "unknown",
                  "cipher": "AES_256_GCM",
                  "issuer": "Let's Encrypt Authority X3",
                  "keyExchange": "ECDHE_RSA",
                  "keyExchangeGroup": "X25519",
                  "protocol": "TLS 1.2",
                  "sanList": [
                    "a1.wpedn.com"
                  ],
                  "signedCertificateTimestampList": [],
                  "subjectName": "a1.wpedn.com",
                  "validFrom": 1577962837,
                  "validTo": 1585738837
                },
                "securityHeaders": [
                  {
                    "name": "Strict-Transport-Security",
                    "value": "max-age=31536000"
                  }
                ],
                "securityState": "secure",
                "status": 200,
                "statusText": "OK",
                "timing": {
                  "connectEnd": -1,
                  "connectStart": -1,
                  "dnsEnd": -1,
                  "dnsStart": -1,
                  "proxyEnd": -1,
                  "proxyStart": -1,
                  "pushEnd": 0,
                  "pushStart": 0,
                  "receiveHeadersEnd": 41.568,
                  "requestTime": 18473661.2033,
                  "sendEnd": 0.283,
                  "sendStart": 0.239,
                  "sslEnd": -1,
                  "sslStart": -1,
                  "workerReady": -1,
                  "workerStart": -1
                },
                "url": "https://a1.wpedn.com/css/fonts.css"
              },
              "size": 2813,
              "type": "Stylesheet"
            }
          },
          {
            "initiatorInfo": {
              "host": "a1.wpedn.com",
              "type": "parser",
              "url": "https://a1.wpedn.com/u"
            },
            "request": {
              "documentURL": "https://a1.wpedn.com/u",
              "frameId": "62F93008F4B77C2C23BAD618EDA45481",
              "hasUserGesture": false,
              "initiator": {
                "lineNumber": 10,
                "type": "parser",
                "url": "https://a1.wpedn.com/u"
              },
              "loaderId": "7041915FE2286C5C0F5B770179C5958B",
              "request": {
                "headers": {
                  "Referer": "https://a1.wpedn.com/u",
                  "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36"
                },
                "initialPriority": "VeryHigh",
                "method": "GET",
                "mixedContentType": "none",
                "referrerPolicy": "no-referrer-when-downgrade",
                "url": "https://a1.wpedn.com/css/site.css?t=2020-01-14T10:18:53Z"
              },
              "requestId": "4996.3",
              "timestamp": 18473661.203244,
              "type": "Stylesheet",
              "wallTime": 1579976485.452952
            },
            "response": {
              "asn": {
                "asn": "16276",
                "country": "FR",
                "date": "2001-02-15",
                "description": "OVH, FR",
                "ip": "91.134.186.165",
                "name": "OVH",
                "registrar": "ripencc",
                "route": "91.134.0.0/16"
              },
              "dataLength": 12823,
              "encodedDataLength": 13078,
              "geoip": {
                "area": 500,
                "city": "",
                "country": "FR",
                "country_name": "France",
                "eu": "1",
                "ll": [
                  48.8582,
                  2.3387
                ],
                "metro": 0,
                "range": [
                  1535556256,
                  1535556271
                ],
                "region": "",
                "timezone": "Europe/Paris"
              },
              "hash": "81dc3432761ae8f618be38c554739ff4489be37564e5eb7e9a5a004cf53af684",
              "hashmatches": [],
              "rdns": {
                "ip": "91.134.186.165",
                "ptr": "rev.kvrmx.com"
              },
              "requestId": "4996.3",
              "response": {
                "encodedDataLength": 255,
                "fromPrefetchCache": false,
                "headers": {
                  "Accept-Ranges": "bytes",
                  "Connection": "keep-alive",
                  "Content-Length": "12823",
                  "Content-Type": "text/css",
                  "Date": "Sat, 25 Jan 2020 18:21:25 GMT",
                  "Last-Modified": "Tue, 14 Jan 2020 10:19:06 GMT",
                  "Server": "nginx",
                  "Strict-Transport-Security": "max-age=31536000"
                },
                "mimeType": "text/css",
                "protocol": "http/1.1",
                "remoteIPAddress": "91.134.186.165",
                "remotePort": 443,
                "securityDetails": {
                  "certificateId": 0,
                  "certificateTransparencyCompliance": "unknown",
                  "cipher": "AES_256_GCM",
                  "issuer": "Let's Encrypt Authority X3",
                  "keyExchange": "ECDHE_RSA",
                  "keyExchangeGroup": "X25519",
                  "protocol": "TLS 1.2",
                  "sanList": [
                    "a1.wpedn.com"
                  ],
                  "signedCertificateTimestampList": [],
                  "subjectName": "a1.wpedn.com",
                  "validFrom": 1577962837,
                  "validTo": 1585738837
                },
                "securityHeaders": [
                  {
                    "name": "Strict-Transport-Security",
                    "value": "max-age=31536000"
                  }
                ],
                "securityState": "secure",
                "status": 200,
                "statusText": "OK",
                "timing": {
                  "connectEnd": -1,
                  "connectStart": -1,
                  "dnsEnd": -1,
                  "dnsStart": -1,
                  "proxyEnd": -1,
                  "proxyStart": -1,
                  "pushEnd": 0,
                  "pushStart": 0,
                  "receiveHeadersEnd": 87.06,
                  "requestTime": 18473661.203403,
                  "sendEnd": 42.121,
                  "sendStart": 42.065,
                  "sslEnd": -1,
                  "sslStart": -1,
                  "workerReady": -1,
                  "workerStart": -1
                },
                "url": "https://a1.wpedn.com/css/site.css?t=2020-01-14T10:18:53Z"
              },
              "size": 12823,
              "type": "Stylesheet"
            }
          },
          {
            "request": {
              "documentURL": "https://a1.wpedn.com/u",
              "frameId": "62F93008F4B77C2C23BAD618EDA45481",
              "hasUserGesture": false,
              "initiator": {
                "type": "other"
              },
              "loaderId": "7041915FE2286C5C0F5B770179C5958B",
              "request": {
                "headers": {
                  "Origin": "https://a1.wpedn.com",
                  "Referer": "https://a1.wpedn.com/css/fonts.css",
                  "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.88 Safari/537.36"
                },
                "initialPriority": "VeryHigh",
…